# docs/swagger.yaml
definitions:
  handler.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  handler.FirstTimePasswordChangeRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
      user_id:
        type: integer
    required:
    - new_password
    - old_password
    - user_id
    type: object
  handler.FirstTimePasswordChangeResponse:
    properties:
      access_token:
        type: string
      message:
        type: string
      roles:
        items:
          type: string
        type: array
      token_type:
        type: string
    type: object
...
  handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.LoginResponse:
    properties:
      access_token:
        type: string
      message:
        type: string
      require_password_change:
        description: For OTP flow
        type: boolean
      token_type:
        type: string
      user:
        $ref: '#/definitions/handler.UserInfo'
      user_id:
        type: integer
    type: object
  handler.UserInfo:
    properties:
      email:
        type: string
      id:
        type: integer
      roles:
        items:
          type: string
        type: array
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorBody'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.ErrorBody:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
info:
  contact: {}
paths:
  /api/v1/change-password:
    post:
      consumes:
      - application/json
      description: Change password for authenticated user
      parameters:
      - description: Password change data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - auth
...
  /auth/first-time-password-change:
    post:
      consumes:
      - application/json
      description: Change one-time password to permanent password
      parameters:
      - description: Password change data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.FirstTimePasswordChangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.FirstTimePasswordChangeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: First-time password change
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and get JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: User login
      tags:
      - auth
...
swagger: "2.0"